Refactorizados US04 y 05 para no acceder directamente a la BD
